# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FarmDirect API",
    "description": "FarmDirect agricultural marketplace API for listing and managing farm produce.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/logistics/serviceability": {
      "get": {
        "summary": "Check pincode serviceability",
        "description": "Delhivery GET /c/api/pin-codes/json/. Call this before booking a shipment.",
        "operationId": "logisticsServiceability",
        "tags": [
          "Logistics"
        ],
        "parameters": [
          {
            "name": "pin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "110033"
          }
        ],
        "responses": {
          "200": {
            "description": "Serviceability result",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "pin": "110033",
                  "serviceable": true,
                  "prepaid": true,
                  "cod": true,
                  "pickup": true,
                  "district": "North Delhi",
                  "stateCode": "DL"
                }
              }
            }
          },
          "400": {
            "description": "pin is not a 6-digit pincode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DELHIVERY_API_KEY is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logistics/create-shipment": {
      "post": {
        "summary": "Create a Delhivery shipment for a matched order",
        "description": "Validates sender/receiver details, then calls Delhivery POST /api/cmu/create.json. On success the shipment (with its AWB) is stored on the order and the order status becomes shipment_booked. Never returns a fabricated success.",
        "operationId": "createShipment",
        "tags": [
          "Logistics"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentRequest"
              },
              "example": {
                "orderId": "ord-1758200000000-123",
                "paymentMode": "Prepaid",
                "delivery": {
                  "name": "Delhi Fresh Mart",
                  "phone": "9876543210",
                  "address": "Shop 4, Azadpur Mandi",
                  "city": "Delhi",
                  "state": "Delhi",
                  "pin": "110033"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shipment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "shipment": {
                      "$ref": "#/components/schemas/Shipment"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; details lists every missing or malformed field",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Shipment data is incomplete or invalid.",
                  "code": "validation_failed",
                  "details": [
                    "delivery.phone must be a valid 10-digit Indian mobile number"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "orderId not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The order already has a shipment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Delhivery rejected the request, was unreachable, or returned an unreadable response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DELHIVERY_API_KEY is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Delhivery timed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logistics/validate-shipment": {
      "post": {
        "summary": "Validate shipment data (no Delhivery call)",
        "description": "Runs the same validation as create-shipment without contacting Delhivery. Useful for pre-flight checks in the UI.",
        "operationId": "validateShipment",
        "tags": [
          "Logistics"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valid",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "valid": true,
                  "errors": []
                }
              }
            }
          },
          "400": {
            "description": "Invalid",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "valid": false,
                  "errors": [
                    "delivery.pin must be a 6-digit pincode"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/logistics/track/{awb}": {
      "get": {
        "summary": "Track a shipment by AWB",
        "description": "Delhivery GET /api/v1/packages/json/. The status is normalised to the app's ShipmentStatus values and written back onto the stored order.",
        "operationId": "trackShipment",
        "tags": [
          "Logistics"
        ],
        "parameters": [
          {
            "name": "awb",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1234512345678"
          }
        ],
        "responses": {
          "200": {
            "description": "Tracking data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed AWB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No shipment found for that AWB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Delhivery error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DELHIVERY_API_KEY is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logistics/cancel-shipment": {
      "post": {
        "summary": "Cancel a shipment",
        "description": "Delhivery POST /api/p/edit with cancellation=true. Allowed while the package is Manifested, In Transit, Pending, Open or Scheduled.",
        "operationId": "cancelShipment",
        "tags": [
          "Logistics"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "awb"
                ],
                "properties": {
                  "awb": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "awb": "1234512345678"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "awb": "1234512345678",
                  "cancelled": true,
                  "remark": "Shipment has been cancelled"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed awb",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Delhivery refused the cancellation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "DELHIVERY_API_KEY is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Validation failed: ..."
          }
        }
      },
...
      "Shipment": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "delhivery"
          },
          "awb": {
            "type": "string",
            "description": "Delhivery waybill number.",
            "example": "1234512345678"
          },
          "orderRef": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "loading",
              "in_transit",
              "delayed",
              "delivered",
              "cancelled"
            ]
          },
          "paymentMode": {
            "type": "string"
          },
          "codAmount": {
            "type": "number"
          },
          "quantityKg": {
            "type": "number"
          },
          "weightGrams": {
            "type": "number"
          },
          "commodity": {
            "type": "string"
          },
          "sortCode": {
            "type": "string",
            "nullable": true
          },
          "pickup": {
            "type": "object"
          },
          "delivery": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LogisticsParty": {
        "type": "object",
        "description": "Sender (farmer/FPO pickup) or receiver (matched buyer) details.",
        "required": [
          "name",
          "phone",
          "address",
          "city",
          "state",
          "pin"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "description": "10-digit Indian mobile; +91/0 prefixes accepted."
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pin": {
            "type": "string",
            "description": "6-digit pincode."
          },
          "country": {
            "type": "string",
            "default": "India"
          },
          "warehouseName": {
            "type": "string",
            "description": "Pickup only. Must exactly match the warehouse registered with Delhivery. Defaults to DELHIVERY_PICKUP_NAME."
          }
        }
      },
      "CreateShipmentRequest": {
        "type": "object",
        "required": [
          "delivery"
        ],
        "properties": {
          "orderId": {
            "type": "string",
            "description": "Id of an order from POST /api/orders. Its commodity, weight and amount are used as the source of truth."
          },
          "requireOrder": {
            "type": "boolean",
            "default": true,
            "description": "Set false to book a shipment without a stored order (commodity and quantityKg must then be supplied)."
          },
          "pickup": {
            "$ref": "#/components/schemas/LogisticsParty"
          },
          "delivery": {
            "$ref": "#/components/schemas/LogisticsParty"
          },
          "commodity": {
            "type": "string"
          },
          "quantityKg": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "paymentMode": {
            "type": "string",
            "enum": [
              "Prepaid",
              "COD"
            ],
            "default": "Prepaid"
          },
          "codAmount": {
            "type": "number",
            "description": "Required and > 0 when paymentMode is COD."
          },
          "shippingMode": {
            "type": "string",
            "enum": [
              "Surface",
              "Express"
            ],
            "default": "Surface"
          },
          "sellerGstTin": {
            "type": "string",
            "description": "Defaults to DELHIVERY_SELLER_GST_TIN."
          },
          "hsnCode": {
            "type": "string",
            "description": "Defaults to DELHIVERY_HSN_CODE."
          }
        }
      },
      "TrackingResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "orderId": {
            "type": "string",
            "nullable": true
          },
          "tracking": {
            "type": "object",
            "properties": {
              "awb": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "scheduled",
                  "loading",
                  "in_transit",
                  "delayed",
                  "delivered",
                  "cancelled"
                ]
              },
              "providerStatus": {
                "type": "string",
                "description": "Raw Delhivery status, e.g. \"In Transit\"."
              },
              "statusLocation": {
                "type": "string",
                "nullable": true
              },
              "statusUpdatedAt": {
                "type": "string",
                "nullable": true
              },
              "origin": {
                "type": "string",
                "nullable": true
              },
              "destination": {
                "type": "string",
                "nullable": true
              },
              "expectedDeliveryAt": {
                "type": "string",
                "nullable": true
              },
              "trackingEvents": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}